+++ test/writer_test.cpp
@@ -312,8 +312,8 @@ TEST_CASE("Writer Pages", "[Writer]")
         REQUIRE(!writer.FindNode(VoxelKey::InvalidKey()).IsValid());
         REQUIRE(!writer.FindNode(VoxelKey(5, 4, 3, 2)).IsValid());
 
-        REQUIRE_NOTHROW(writer.GetPage());
-        Page root_page = writer.GetPage();
+        REQUIRE_NOTHROW(writer.GetRootPage());
+        Page root_page = writer.GetRootPage();
         REQUIRE(root_page.IsValid());
         REQUIRE(root_page.IsPage());
         REQUIRE(root_page.loaded == true);
@@ -334,12 +334,20 @@ TEST_CASE("Writer Pages", "[Writer]")
 
         Writer writer(out_stream, {6});
 
-        Page root_page = writer.GetPage();
+        Page root_page = writer.GetRootPage();
 
         auto sub_page = writer.AddSubPage(root_page, VoxelKey(1, 1, 1, 1));
         REQUIRE(sub_page.IsPage());
         REQUIRE(sub_page.IsValid());
         REQUIRE(sub_page.loaded == true);
+
+        // GetPage Test
+        auto found_page = writer.GetPage(VoxelKey(1, 1, 1, 1));
+        REQUIRE(found_page.offset == sub_page.offset);
+        REQUIRE(found_page.byte_size == sub_page.byte_size);
+        REQUIRE(found_page.point_count == sub_page.point_count);
+
+        REQUIRE_THROWS(writer.GetPage(VoxelKey(2, 2, 2, 2)));
 
         REQUIRE_THROWS(writer.AddSubPage(sub_page, VoxelKey(1, 1, 1, 0)));
         REQUIRE_THROWS(writer.AddSubPage(sub_page, VoxelKey(2, 4, 5, 0)));
@@ -446,7 +454,7 @@ TEST_CASE("Writer Copy", "[Writer]")
 
         Writer writer(out_stream, cfg);
 
-        Page root_page = writer.GetPage();
+        Page root_page = writer.GetRootPage();
 
         for (const auto &node : reader.GetAllNodes())
         {
@@ -488,7 +496,7 @@ TEST_CASE("Check Spatial Bounds", "[Writer]")
         FileWriter writer(file_path, cfg);
 
         auto header = *writer.CopcConfig()->LasHeader();
-        Page root_page = writer.GetPage();
+        Page root_page = writer.GetRootPage();
 
         las::Points points(header.PointFormatId(), header.Scale(), header.Offset());
 
@@ -511,7 +519,7 @@ TEST_CASE("Check Spatial Bounds", "[Writer]")
         FileWriter writer(file_path, cfg);
 
         auto header = *writer.CopcConfig()->LasHeader();
-        Page root_page = writer.GetPage();
+        Page root_page = writer.GetRootPage();
 
         las::Points points(header.PointFormatId(), header.Scale(), header.Offset());
 
@@ -534,7 +542,7 @@ TEST_CASE("Check Spatial Bounds", "[Writer]")
         FileWriter writer(file_path, cfg);
 
         auto header = *writer.CopcConfig()->LasHeader();
-        Page root_page = writer.GetPage();
+        Page root_page = writer.GetRootPage();
 
         las::Points points(header.PointFormatId(), header.Scale(), header.Offset());
 
@@ -557,7 +565,7 @@ TEST_CASE("Check Spatial Bounds", "[Writer]")
         FileWriter writer(file_path, cfg);
 
         auto header = *writer.CopcConfig()->LasHeader();
-        Page root_page = writer.GetPage();
+        Page root_page = writer.GetRootPage();
 
         las::Points points(header.PointFormatId(), header.Scale(), header.Offset());