+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- **\[Python/C++\]** Added `GetPage(VoxelKey key)` to `Writer`
+
 ### Changed
 
 - **\[Python/C++\]** Write WKT as EVLR instead of VLR to align with laspy
 - **\[Python/C++\]** Removed the error when scan angle is not within LAS specs [-30000,30000]
-- **\[Python/C++\]** Modified `Writer`'s `GetRootPage()` to `GetPage(VoxelKey key = VoxelKey::RootKey())`
 
 ## [2.0.0] - 2021-10-28